+++ src/PSRule/Commands/AssertExistsCommand.cs
@@ -9,32 +9,55 @@ namespace PSRule.Commands
     [Cmdlet(VerbsLifecycle.Assert, RuleLanguageNouns.Exists)]
     internal sealed class AssertExistsCommand : RuleKeyword
     {
+        public AssertExistsCommand()
+        {
+            CaseSensitive = false;
+            Not = false;
+        }
+
         [Parameter(Mandatory = true, Position = 0)]
         public string[] Field { get; set; }
 
+        [Parameter(Mandatory = false)]
+        [PSDefaultValue(Value = false)]
+        public SwitchParameter CaseSensitive { get; set; }
+
+        [Parameter(Mandatory = false)]
+        [PSDefaultValue(Value = false)]
+        public SwitchParameter Not { get; set; }
+
         protected override void ProcessRecord()
         {
-            PipelineContext.WriteVerbose("[Exists]::BEGIN");
-
             try
             {
                 var inputObject = GetVariableValue("InputObject") ?? GetVariableValue("TargetObject");
 
-                bool result = false;
+                bool expected = !Not;
+                bool actual = Not;
 
-                foreach (var fieldName in Field)
+                for (var i = 0; i < Field.Length && actual != expected; i++)
                 {
-                    if (GetField(inputObject, fieldName, out object fieldValue))
+                    actual = GetField(inputObject, Field[i], CaseSensitive, out object fieldValue);
+
+                    if (actual == expected)
                     {
-                        result = true;
+                        if (expected)
+                        {
+                            PipelineContext.WriteVerbose($"[Exists] -- The field {Field[i]} exists");
+                        }
                     }
                 }
 
-                WriteObject(result);
+                if (!actual)
+                {
+                    PipelineContext.WriteVerbose($"[Exists] -- The field(s) {string.Join(", ", Field)} do not exist");
+                }
+
+                WriteObject(expected == actual);
             }
             finally
             {
-                PipelineContext.WriteVerbose("[Exists]::END");
+                
             }
         }
     }