+++ src/PSRule/Pipeline/PipelineContext.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Management.Automation;
+using PSRule.Host;
+using PSRule.Rules;
 
 namespace PSRule.Pipeline
 {
@@ -8,10 +10,13 @@ namespace PSRule.Pipeline
         [ThreadStatic]
         internal static PipelineContext CurrentThread;
 
+        private string _LogPrefix;
+        private int _ObjectNumber;
         private readonly ILogger _Logger;
 
         private PipelineContext(ILogger logger)
         {
+            _ObjectNumber = 0;
             _Logger = logger;
         }
 
@@ -27,9 +32,9 @@ namespace PSRule.Pipeline
             CurrentThread.DoWriteError(errorRecord);
         }
 
-        public static void WriteVerbose(string message)
+        public static void WriteVerbose(string message, bool usePrefix = true)
         {
-            CurrentThread.DoWriteVerbose(message);
+            CurrentThread.DoWriteVerbose(message, usePrefix);
         }
 
         public static void WriteWarning(string message)
@@ -47,15 +52,30 @@ namespace PSRule.Pipeline
             _Logger.WriteError(errorRecord);
         }
 
-        private void DoWriteVerbose(string message)
+        private void DoWriteVerbose(string message, bool usePrefix)
         {
             if (_Logger == null)
             {
                 return;
             }
 
+            var outMessage = message;
+
+            if (usePrefix)
+            {
+                outMessage = string.Concat(_LogPrefix, message);
+            }
+
             // TODO: Prefix entries with context. i.e. //[Rule][$($Context.Index)][$($Rule.Name)]
-            _Logger.WriteVerbose(message);
+            _Logger.WriteVerbose(outMessage);
+        }
+
+        /// <summary>
+        /// Increment the pipeline object number.
+        /// </summary>
+        public void Next()
+        {
+            _ObjectNumber++;
         }
 
         private void DoWriteWarning(string message)
@@ -66,6 +86,16 @@ namespace PSRule.Pipeline
             }
 
             _Logger.WriteVerbose(message);
+        }
+
+        public void Enter(DependencyGraph<RuleBlock>.DependencyTarget target)
+        {
+            _LogPrefix = $"[PSRule][R][{_ObjectNumber}][{target.Value.RuleId}]";
+        }
+
+        public void Exit()
+        {
+            _LogPrefix = string.Empty;
         }
     }
 }