+++ src/PSRule/Rules/Rule.cs
@@ -9,13 +9,29 @@ namespace PSRule.Rules
     [JsonObject]
     public sealed class Rule
     {
+        /// <summary>
+        /// A unique identifier for the rule.
+        /// </summary>
         [JsonProperty(PropertyName = "ruleId", Required = Required.Always)]
         public string RuleId { get; set; }
 
+        /// <summary>
+        /// The name of the rule.
+        /// </summary>
+        [JsonProperty(PropertyName = "ruleName", Required = Required.Always)]
+        public string RuleName { get; set; }
+
+        /// <summary>
+        /// The script file path where the rule is defined.
+        /// </summary>
         [JsonProperty(PropertyName = "sourcePath")]
         public string SourcePath { get; set; }
 
+        /// <summary>
+        /// A human readable block of text, used to identify the purpose of the rule.
+        /// </summary>
         [JsonProperty(PropertyName = "description")]
+        [DefaultValue(null)]
         public string Description { get; set; }
 
         /// <summary>