+++ src/PSRule/Rules/RuleFilter.cs
@@ -9,17 +9,17 @@ namespace PSRule.Rules
     /// </summary>
     public sealed class RuleFilter
     {
-        private HashSet<string> _RequiredRuleId;
+        private HashSet<string> _RequiredRuleName;
         private Hashtable _RequiredTag;
 
         /// <summary>
         /// Filter rules by id or tag.
         /// </summary>
-        /// <param name="ruleId"></param>
+        /// <param name="ruleName"></param>
         /// <param name="tag"></param>
-        public RuleFilter(IEnumerable<string> ruleId, Hashtable tag)
+        public RuleFilter(IEnumerable<string> ruleName, Hashtable tag)
         {
-            _RequiredRuleId = ruleId == null ? null : new HashSet<string>(ruleId, StringComparer.OrdinalIgnoreCase);
+            _RequiredRuleName = ruleName == null ? null : new HashSet<string>(ruleName, StringComparer.OrdinalIgnoreCase);
             _RequiredTag = tag ?? null;
         }
 
@@ -27,9 +27,9 @@ namespace PSRule.Rules
         /// Matches if the RuleId is contained or any tag is matched
         /// </summary>
         /// <returns>Return true if rule is matched, otherwise false.</returns>
-        public bool Match(string ruleId, TagSet tag)
+        public bool Match(string ruleName, TagSet tag)
         {
-            if (_RequiredRuleId == null || _RequiredRuleId.Contains(ruleId))
+            if (_RequiredRuleName == null || _RequiredRuleName.Contains(ruleName))
             {
                 if (_RequiredTag == null)
                 {
@@ -57,7 +57,7 @@ namespace PSRule.Rules
 
         public bool Match(RuleBlock block)
         {
-            return Match(block.RuleId, block.Tag);
+            return Match(block.RuleName, block.Tag);
         }
     }
 }